requirements.txt
numpy==1.19.5
opencv-contrib-python==4.5.3.56
matplotlib==3.4.3
pillow
glob2
pathlib